feat: add note color

+++ app/src/main/java/com/github/tamal8730/noteit/core/data/entity/NoteEntity.kt
@@ -19,6 +19,6 @@ data class NoteEntity(
     val coverImage: String?,
     val lastModifiedAt: String,
 
-    val color: Int?
+    val color: Long?
 
 )
