+++ app/src/main/java/com/github/tamal8730/noteit/feature_arrange_notes/view/notes_grid_screen/ui_model/NoteUIModel.kt
@@ -9,5 +9,5 @@ data class NoteUIModel(
     val coverImage: Uri?,
     val lastModifiedAt: String,
     val tasks: List<TaskUIModel>?,
-    val color: Int?,
+    val color: Long?,
 )
